Optimize memory: use UInt16 for positions and Int16 for scores
Changes:
- MatchResult.score: Int → Int16 (scores rarely exceed 32K)
- MatchResult.positions: [Int] → [UInt16] (line lengths up to 65,535 bytes)
- Updated FuzzyMatcher, Algorithm, and Utf8FuzzyMatch to build UInt16 arrays directly
- Updated TextRenderer functions to accept [UInt16]
- Fixed all tests to use UInt16 array literals

Memory savings:
- Before: MatchedItem = 40 bytes (stack) + ~72 bytes (heap for 5 positions with [Int])
  = 112 bytes total
- After: MatchedItem = 40 bytes (stack) + ~42 bytes (heap for 5 positions with [UInt16])
  = 82 bytes total
- Savings: 30 bytes per MatchedItem (27% reduction)

For 400k matched items: 400,000 × 30 = 12 MB saved
For 50% selectivity on find.txt (413k matches): ~12.4 MB saved
For 90% selectivity on find.txt (745k matches): ~22.4 MB saved

Expected impact: Reduces memory from 243 MB → ~220-230 MB

+++ Sources/FltrLib/Matcher/Utf8FuzzyMatch.swift
@@ -420,7 +420,7 @@ public struct Utf8FuzzyMatch: Sendable {
         guard bestScore > 0 else { return nil }
 
         // ── Phase 4: backtrack ─────────────────────────────────────────
-        var positions = [Int](repeating: 0, count: M)
+        var positions = [UInt16](repeating: 0, count: M)
 
         buf.I16.withUnsafeBufferPointer { i16buf in
             buf.I32.withUnsafeBufferPointer { i32buf in
@@ -450,7 +450,7 @@ public struct Utf8FuzzyMatch: Sendable {
                         }
 
                         if s > s1 && (s > s2 || (s == s2 && preferMatch)) {
-                            pos[i] = minIdx + j
+                            pos[i] = UInt16(minIdx + j)
                             if i == 0 { break }
                             i -= 1
                         }
@@ -465,6 +465,6 @@ public struct Utf8FuzzyMatch: Sendable {
             }
         }
 
-        return MatchResult(score: Int(bestScore), positions: positions)
+        return MatchResult(score: Int16(bestScore), positions: positions)
     }
 }
